+++ parboiled-core/src/main/scala/org/parboiled2/Rule.scala
@@ -20,7 +20,7 @@ import scala.annotation.unchecked.uncheckedVariance
 import scala.annotation.compileTimeOnly
 import scala.collection.immutable
 import org.parboiled2.support._
-import shapeless.HList
+import org.parboiled2.support.hlist.HList
 
 sealed trait RuleX
