+++ parboiled-core/src/main/scala/org/parboiled2/RuleDSLBasics.scala
@@ -18,7 +18,7 @@ package org.parboiled2
 
 import scala.annotation.compileTimeOnly
 import org.parboiled2.support._
-import shapeless.HList
+import org.parboiled2.support.hlist.HList
 
 trait RuleDSLBasics {
