+++ parboiled-core/src/main/scala/org/parboiled2/support/ActionOpsSupport.scala
@@ -16,9 +16,9 @@
 
 package org.parboiled2.support
 
-import shapeless._
+import org.parboiled2.support.hlist._
 import org.parboiled2.Rule
-import shapeless.ops.hlist.Prepend
+import org.parboiled2.support.hlist.ops.hlist.Prepend
 
 /*
  * The main ActionOps boilerplate is generated by a custom SBT sourceGenerator.
