+++ parboiled-core/src/main/scala/org/parboiled2/support/Unpack.scala
@@ -16,7 +16,7 @@
 
 package org.parboiled2.support
 
-import shapeless._
+import org.parboiled2.support.hlist._
 
 /** "Unpacks" an HList if it has only zero or one element(s).
   *   Out =
