+++ parboiled-core/src/test/scala/org/parboiled2/ActionSpec.scala
@@ -16,7 +16,7 @@
 
 package org.parboiled2
 
-import shapeless._
+import org.parboiled2.support.hlist._
 import utest._
 
 object ActionSpec extends TestParserSpec {
