+++ parboiled-core/src/test/scala/org/parboiled2/VarianceSpec.scala
@@ -16,7 +16,7 @@
 
 package org.parboiled2
 
-import shapeless._
+import org.parboiled2.support.hlist._
 import utest.{compileError, ArrowAssert}
 
 //// pure compile-time-only test
